[IMP] web: convert scroller tests to hoot task-id: 3705027
